[C-1205] Improve splash screen animation
