+++ packages/mobile/src/screens/root-screen/RootScreen.tsx
@@ -1,31 +1,21 @@
 import { useEffect, useState } from 'react'
 
 import { accountSelectors, Status } from '@audius/common'
-import type { DrawerContentComponentProps } from '@react-navigation/drawer'
-import { createDrawerNavigator } from '@react-navigation/drawer'
-// eslint-disable-next-line import/no-unresolved
-import type { DrawerNavigationHelpers } from '@react-navigation/drawer/lib/typescript/src/types'
 import type { NavigatorScreenParams } from '@react-navigation/native'
-import { useNavigation } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import { setupBackend } from 'audius-client/src/common/store/backend/actions'
-import { Dimensions } from 'react-native'
 import { useDispatch, useSelector } from 'react-redux'
 
 import useAppState from 'app/hooks/useAppState'
 import { useUpdateRequired } from 'app/hooks/useUpdateRequired'
-import PushNotifications from 'app/notifications'
 import type { AppScreenParamList } from 'app/screens/app-screen'
-import { AppScreen } from 'app/screens/app-screen'
-import {
-  NotificationsScreen,
-  NotificationsDrawerNavigationContextProvider
-} from 'app/screens/notifications-screen'
 import { SignOnScreen } from 'app/screens/signon'
 import { UpdateRequiredScreen } from 'app/screens/update-required-screen/UpdateRequiredScreen'
 import { enterBackground, enterForeground } from 'app/store/lifecycle/actions'
 
 import { SplashScreen } from '../splash-screen'
+
+import { HomeScreen } from './HomeScreen'
 
 const { getHasAccount, getAccountStatus } = accountSelectors
 
@@ -36,96 +26,7 @@ export type RootScreenParamList = {
   }>
 }
 
-const SCREEN_WIDTH = Dimensions.get('window').width
-
-const Drawer = createDrawerNavigator()
 const Stack = createNativeStackNavigator()
-
-/**
- * The sign up & sign in stack when not authenticated
- */
-const SignOnStack = () => {
-  return (
-    <Stack.Navigator
-      screenOptions={{ gestureEnabled: false, headerShown: false }}
-    >
-      <Stack.Screen name='SignOnStack' component={SignOnScreen} />
-    </Stack.Navigator>
-  )
-}
-
-/**
- * Update stack when the app is behind the minimum app version
- */
-const UpdateStack = () => {
-  return (
-    <Stack.Navigator
-      screenOptions={{ gestureEnabled: false, headerShown: false }}
-    >
-      <Stack.Screen name='UpdateStack' component={UpdateRequiredScreen} />
-    </Stack.Navigator>
-  )
-}
-
-type MainStackProps = {
-  navigation: DrawerNavigationHelpers
-}
-
-/**
- * The main stack after signing up or signing in
- */
-const MainStack = (props: MainStackProps) => {
-  const { navigation: drawerHelpers } = props
-  const drawerNavigation = useNavigation()
-  return (
-    <NotificationsDrawerNavigationContextProvider
-      drawerNavigation={drawerNavigation}
-      drawerHelpers={drawerHelpers}
-    >
-      <Stack.Navigator
-        screenOptions={{ gestureEnabled: false, headerShown: false }}
-      >
-        <Stack.Screen name='MainStack' component={AppScreen} />
-      </Stack.Navigator>
-    </NotificationsDrawerNavigationContextProvider>
-  )
-}
-
-type NotificationDrawerContentsProps = DrawerContentComponentProps & {
-  disableGestures: boolean
-  setDisableGestures: (disabled: boolean) => void
-}
-
-/**
- * The contents of the notifications drawer, which swipes in
- */
-const NotificationsDrawerContents = (
-  props: NotificationDrawerContentsProps
-) => {
-  const {
-    navigation: drawerHelpers,
-    disableGestures,
-    setDisableGestures,
-    state
-  } = props
-  const drawerNavigation = useNavigation()
-
-  useEffect(() => {
-    PushNotifications.setDrawerHelpers(drawerHelpers)
-  }, [drawerHelpers])
-
-  return (
-    <NotificationsDrawerNavigationContextProvider
-      drawerHelpers={drawerHelpers}
-      drawerNavigation={drawerNavigation}
-      gesturesDisabled={disableGestures}
-      setGesturesDisabled={setDisableGestures}
-      state={state}
-    >
-      <NotificationsScreen />
-    </NotificationsDrawerNavigationContextProvider>
-  )
-}
 
 type RootScreenProps = {
   isReadyToSetupBackend: boolean
@@ -139,7 +40,7 @@ export const RootScreen = ({ isReadyToSetupBackend }: RootScreenProps) => {
   const dispatch = useDispatch()
   const hasAccount = useSelector(getHasAccount)
   const accountStatus = useSelector(getAccountStatus)
-  const [disableGestures, setDisableGestures] = useState(false)
+  const [isInitting, setIsInittng] = useState(true)
   const { updateRequired } = useUpdateRequired()
 
   useEffect(() => {
@@ -154,41 +55,27 @@ export const RootScreen = ({ isReadyToSetupBackend }: RootScreenProps) => {
     () => dispatch(enterBackground())
   )
 
-  const isAccountLoading =
-    accountStatus === Status.IDLE || accountStatus === Status.LOADING
+  useEffect(() => {
+    if (accountStatus === Status.SUCCESS || accountStatus === Status.ERROR) {
+      setIsInittng(false)
+    }
+  }, [accountStatus])
 
   return (
     <>
       <SplashScreen />
-      {updateRequired ? <UpdateStack /> : null}
-      {isAccountLoading ? null : !hasAccount ? (
-        <SignOnStack />
-      ) : (
-        <Drawer.Navigator
-          // legacy implementation uses reanimated-v1
-          useLegacyImplementation={true}
-          detachInactiveScreens={false}
-          screenOptions={{
-            drawerType: 'slide',
-            headerShown: false,
-            drawerStyle: {
-              width: '100%'
-            },
-            swipeEdgeWidth: SCREEN_WIDTH,
-            gestureHandlerProps: {
-              enabled: !disableGestures
-            }
-          }}
-          drawerContent={(props) => (
-            <NotificationsDrawerContents
-              disableGestures={disableGestures}
-              setDisableGestures={setDisableGestures}
-              {...props}
-            />
-          )}
+      {isInitting && !hasAccount ? null : (
+        <Stack.Navigator
+          screenOptions={{ gestureEnabled: false, headerShown: false }}
         >
-          <Drawer.Screen name='App' component={MainStack} />
-        </Drawer.Navigator>
+          {updateRequired ? (
+            <Stack.Screen name='UpdateStack' component={UpdateRequiredScreen} />
+          ) : !hasAccount ? (
+            <Stack.Screen name='SignOnStack' component={SignOnScreen} />
+          ) : (
+            <Stack.Screen name='DrawerStack' component={HomeScreen} />
+          )}
+        </Stack.Navigator>
       )}
     </>
   )
